Move '/api' into separate file
